+++ scrapy/core/downloader/contextfactory.py
@@ -22,14 +22,14 @@ from scrapy.core.downloader.tls import (
 )
 from scrapy.exceptions import ScrapyDeprecationWarning
 from scrapy.utils._deps_compat import TWISTED_TLS_NEW_IMPL
-from scrapy.utils.deprecate import create_deprecated_class
-from scrapy.utils.misc import build_from_crawler, load_object
-from scrapy.utils.ssl import (
+from scrapy.utils._ssl import (
     _get_cert_options_version_kwargs,
     _get_keylog_filename,
     _get_tls_version_limits,
     _set_keylog_callback,
 )
+from scrapy.utils.deprecate import create_deprecated_class
+from scrapy.utils.misc import build_from_crawler, load_object
 
 if TYPE_CHECKING:
     from twisted.internet._sslverify import ClientTLSOptions
