+++ tests/ignores.txt
@@ -5,4 +5,5 @@ scrapy/mail.py
 scrapy/utils/console.py
 scrapy/utils/gz.py
 scrapy/utils/job.py
+scrapy/utils/ssl.py
 scrapy/utils/versions.py
